lineup-parameters: do not insert tabs vs spaces dynamically, add --tabs
It was a bad idea to decide whether to insert spaces or tabs dynamically
by looking at the initial content.

For example with the GNU coding style, the indentation/alignment is
always with spaces. If a contributor inserts by mistake a tab to align a
parameter, the script should do the right thing and enforce alignment
with spaces.

So now by default spaces are inserted, and there is a --tabs command
line option to force tabs+spaces.

Why spaces by default? Because GLib/GTK+ use the GNU coding style. And
for gnome-c-utils, GLib/GTK+ is the reference. Also for some projects,
the preferred coding style is tabs for indentation, but spaces for
*alignment*.