meson: simplify code to compute whether all Tepl deps are found
The Meson intro manual explained that all variables are immutable, but
it's actually not the case, it is possible to reassign a new value to a
variable, what is immutable is the *object* stored inside the variable.

The new code has actually the same number of lines (I've removed the
comment so the diff shows two lines less), but it is more usual code,
with one less variable.